feat(CA-174): consolidate ProjectSettingsBloc and wire create-project route
- Move full BLoC (events, states, handler) to lib/libraries/project/bloc/
  so it's accessible to all features; dashboard re-exports kept as thin aliases
- Add ProjectSettingsCreationRequested handler with null-user guard
- Add project_settings_routes.dart constants and wire ProjectSettingsRoutesModule
  into ShellModule
- Wire bottom sheet "Create project" to push createProjectRoute
