+++ lib/libraries/router/routes/project_settings_routes.dart
@@ -1,0 +1,3 @@
+const String projectSettingsBaseRoute = '/project-settings';
+const String createProjectChildRoute = '/create-project';
+const String createProjectRoute = '$projectSettingsBaseRoute$createProjectChildRoute';
